fixed: all Animations are using the same timer now. fixed: small rounding difference when showing FPS

+++ mediaportal/xbmc/d3dapp.cs
@@ -1878,7 +1878,7 @@ namespace MediaPortal
       // Update the scene stats once per second
       if (time - lastTime >= 1.0f)
       {
-        framePerSecond = frames / (time - lastTime);
+        framePerSecond = (float)frames / (time - lastTime);
         GUIGraphicsContext.CurrentFPS = framePerSecond;
         lastTime = time;
         frames = 0;
